src/BackendFastApi/ocrProjectBackend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
langdetect==1.0.9
python-docx==1.1.0
reportlab==4.0.7
paddlepaddle==2.6.0
paddleocr==2.7.3
